feat(google-calendar): automatic sync via scheduled cron
- Extract sync logic into src/lib/google-calendar-sync.js (shared)
- Refactor per-user sync route to use the shared lib
- Add /api/google-calendar/sync-all route (CRON_SECRET auth) that syncs
  all connected users; per-connection failures are isolated and logged
- Add calendar-sync-cron Netlify scheduled function (every 30 min)
- Register schedule in netlify.toml; document automatic sync in setup guide

Closes the gap where jobs only synced on manual 'Sync Now' click.
