+++ netlify/functions/calendar-sync-cron.ts
@@ -1,0 +1,32 @@
+// Netlify Scheduled Function: Google Calendar Sync
+// Runs every 30 minutes via netlify.toml schedule.
+// Pushes each connected user's upcoming jobs to their Google Calendar so the
+// sync is automatic — users no longer need to click "Sync Now" in Settings.
+
+export default async function handler() {
+  const siteUrl = process.env.URL || process.env.DEPLOY_PRIME_URL || 'http://localhost:3000';
+  const cronSecret = process.env.CRON_SECRET || '';
+
+  try {
+    const response = await fetch(`${siteUrl}/api/google-calendar/sync-all`, {
+      method: 'GET',
+      headers: {
+        'Authorization': `Bearer ${cronSecret}`,
+      },
+    });
+
+    const data = await response.json();
+    console.log('[Calendar Sync Cron] Results:', JSON.stringify(data));
+
+    return new Response(JSON.stringify({ success: true, ...data }), {
+      status: 200,
+      headers: { 'Content-Type': 'application/json' },
+    });
+  } catch (error) {
+    console.error('[Calendar Sync Cron] Error:', error);
+    return new Response(JSON.stringify({ error: 'Cron execution failed' }), {
+      status: 500,
+      headers: { 'Content-Type': 'application/json' },
+    });
+  }
+}
